add modules and class
