+++ lib/check_game_result.rb
@@ -1,28 +1,9 @@
+require_relative './win_vars'
 class CheckGameResult
   def initialize(got_board)
     @board = got_board
   end
-
-  def win_case
-    @win_cases = [
-      [0, 1, 2],
-      [3, 4, 5],
-      [6, 7, 8],
-      [0, 4, 8],
-      [2, 4, 6],
-      [2, 5, 8],
-      [1, 4, 7],
-      [0, 3, 6]
-    ]
-  end
-
-  def win_var
-    @count_win_x = 0
-    @count_win_o = 0
-    @every_move = 0
-    @win_leng = @win_cases.length
-  end
-
+  include WinVar
   def loop_on_cases
     @win_cases[@every_move].each do |item|
       @count_win_x += 1 if @board[item] == 'X'
